+++ ruber FIX/.gitignore
@@ -1,0 +1,2 @@
+ruber465.VC.VC.opendb
+Debug/*
