+++ ruber FIX/ruber465.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ruber465", "ruber465.vcxproj", "{12087D28-1F55-43F5-84FE-AB2C584607BA}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{12087D28-1F55-43F5-84FE-AB2C584607BA}.Debug|x64.ActiveCfg = Debug|x64
+		{12087D28-1F55-43F5-84FE-AB2C584607BA}.Debug|x64.Build.0 = Debug|x64
+		{12087D28-1F55-43F5-84FE-AB2C584607BA}.Debug|x86.ActiveCfg = Debug|Win32
+		{12087D28-1F55-43F5-84FE-AB2C584607BA}.Debug|x86.Build.0 = Debug|Win32
+		{12087D28-1F55-43F5-84FE-AB2C584607BA}.Release|x64.ActiveCfg = Release|x64
+		{12087D28-1F55-43F5-84FE-AB2C584607BA}.Release|x64.Build.0 = Release|x64
+		{12087D28-1F55-43F5-84FE-AB2C584607BA}.Release|x86.ActiveCfg = Release|Win32
+		{12087D28-1F55-43F5-84FE-AB2C584607BA}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
